<Feat> : 테이블명 변경 및 서버리스트 entity 추가

+++ hamonize-center/src/main/java/com/mapper/IActAgentLogInOutMapper.java
@@ -8,5 +8,5 @@ public interface IActAgentLogInOutMapper {
 	public int insertLoginLog(LogInOutVo vo);
 	public int updateLoginLog(LogInOutVo vo);
 	
-	public int selectLoginLogSeq(LogInOutVo vo);
+	public Long selectLoginLogSeq(LogInOutVo vo);
 }
